Fix: Delete버튼 클릭 시, 화면에 바로 반영
새로고침해야 document 항목을 삭제한 것이 반영되는 문제 수정

+++ src/PostList.js
@@ -29,18 +29,30 @@ export default function PostList({ $target, initialState }) {
 
   //   this.render();
 
-  $postList.addEventListener("click", (e) => {
+  $postList.addEventListener("click", async (e) => {
     const $li = e.target.closest("li"); // postList내 클릭한 li 가져오기
-    const { id } = $li.dataset;
-    if ($li) {
-      push(`/documents/${id}`);
-    }
-
+    const $buttonAdd = e.target.closest("li > button#add");
     const $buttonDelete = e.target.closest("li > button#delete");
-    if ($buttonDelete) {
-      request(`/documents/${id}`, {
-        method: "DELETE",
-      });
+
+    const { id } = $li.dataset;
+
+    if ($li && $buttonAdd) {
+      push(`/documents/new`);
+    } else if ($li && $buttonDelete) {
+      try {
+        await request(`/documents/${id}`, {
+          method: "DELETE",
+        });
+
+        // 서버 요청이 성공하면 클라이언트 상태 업데이트
+        this.state = this.state.filter((post) => post.id !== Number(id));
+
+        this.render();
+      } catch (error) {
+        console.error("삭제 중 오류 발생:", error);
+      }
+    } else {
+      push(`/documents/${id}`);
     }
   });
 }
